+++ src/widgets/qownnotesmarkdowntextedit.h
@@ -17,6 +17,13 @@ public:
     };
     Q_ENUMS(EditorWidthMode)
 
+    enum FontModificationMode {
+        Increase = 1,
+        Decrease,
+        Reset
+    };
+    Q_ENUMS(FontModificationMode)
+
     explicit QOwnNotesMarkdownTextEdit(QWidget *parent = 0);
 
     void setStyles();
@@ -24,6 +31,7 @@ public:
 //    void setViewportMargins(int left, int top, int right, int bottom);
     void setPaperMargins(int width);
     void setMainWindow(MainWindow *mainWindow);
+    int modifyFontSize(FontModificationMode mode);
 
 protected:
     void insertFromMimeData(const QMimeData * source);